requirements.txt
streamlit>=1.35.0
openai>=1.30.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0